config-loader: move transform/ into sources/
